New: Remove goal from IRunOptions; Pass directly to runners (fixes #7)
Hide kanren underneath a builder. This is the first step to provided option to build a specialized kanren with custom constraints and term types. (refs #2) (#8)
The IRunOptions type was modified; the goal property was removed. Instead, goals will be passed as unnamed arguments to runners to make invocations of kanren more concise. (refs #7)